update README and move useEffect to custom hook
